New version of deployment scripts for Filesender

+++ deploy_full_procedure.sh
@@ -51,7 +51,13 @@ fi
 sudo yum install git
 
 # Needed  to uglify the js files
-sudo yum install npm.x86_64
+
+NODE_EXISTS=$(which node)
+if [ $NODE_EXISTS  ]; then
+	echo "Node/npm is installed and run from $NODE_EXISTS"
+else
+	sudo yum install npm.x86_64
+fi
 
 ## Check it /work is a mounted directory
 if [ "${workonabel}" == "y" ]; then
@@ -113,22 +119,28 @@ if [ "${GALAXY_ABEL_MOUNT}" == "1" ]; then
 		
 		cd ${MYDIR}/filesender_setup
 		sudo sh -c "./deploy_filesender_root.sh"
+
+		## fix the selinux context for filesender and simplesaml
+		sudo semanage fcontext -a -t httpd_sys_content_t -s system_u '/opt/filesender(/.*)?'
+		sudo semanage fcontext -a -t httpd_sys_rw_content_t -s system_u '/opt/filesender/filesender/(log|tmp|files)(/.*)?'
+		sudo semanage fcontext -a -t httpd_sys_rw_content_t -s system_u '/opt/filesender/simplesaml/log(/.*)?'
+		sudo restorecon -FR /opt/filesender
 		
 		# edit Galaxy config files
 		sudo -u ${GALAXYUSER} -H sh -c "./deploy_filesender_galaxy_config.sh"
 
-
 		# Last instructions:
-		echo -e "\n==== LAST INSTRUCTIONS FOR Filesender SETUP ==== "
-		echo -e "The filesender storage directory must belong to 'apache' user (or nobody) and be writable for group 'galaxy'\n"
-		echo "Log into nh.abel as root, cd to ABEL_FILESENDER_PATH (e.g. /work/projects/galaxy/filesender) and run :"
+		echo -e "\n==== LAST INSTRUCTIONS FOR Filesender SETUP ==== \n"
+		## The filesender storage and simplesaml logs directory must belong to 'apache' user (or nobody) and be writable for group 'galaxy'
+		echo "Log into nh.abel as root"
+		echo "cd to ABEL_FILESENDER_PATH (e.g. /work/projects/galaxy/filesender) and run :"
 		echo "chown -R apache *"
 		echo "chmod -R g+w *"
-		echo "chmod a+w ${GALAXY_PUBLIC_HOSTNAME}/ssl_error_log"
-		echo "chmod a+w ${GALAXY_PUBLIC_HOSTNAME}/ssl_access_log"
+                echo "cd to ABEL_SIMPLESAML_PATH (e.g. /work/projects/galaxy/simplesaml) and run :"
+                echo "chown -R apache *"
+                echo "chmod -R g+w *"
 		echo -e "\n==== Filesender setup READY! ====\n\n"
 
-				
 		# Get back to the main level
 		cd ${MYDIR}
     fi
@@ -150,5 +162,3 @@ if [[ "${GALAXY_ABEL_MOUNT}" == "1" ]]; then
 else
     cat ${MYDIR}/POST_INSTALLATION_INDEPENDENT.md
 fi
-
-
